update tablet and iphone5 ver and refactor
