+++ pages/index.tsx
@@ -21,11 +21,11 @@ export default function Home() {
         </header>
 
         <div className={styles.iphonePic}>
-          <Image src={iphone12} alt="iPhone12" />
+          <Image src={iphone12} alt="iPhone12" placeholder="blur" />
         </div>
 
         <div className={styles.laptopPic}>
-          <Image src={blackDesktop} alt="Black Desktop" />
+          <Image src={blackDesktop} alt="Black Desktop" placeholder="blur" />
         </div>
 
         <div style={{ zIndex: 4 }}>
@@ -34,11 +34,11 @@ export default function Home() {
       </div>
 
       <div>
-        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 1440 1600">
+        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 1440 1700">
           <path
             fill="#000a33"
             fillOpacity="1"
-            d="M0,1600 L0,370 L1440,0 L1440,1600 L0,1600 Z"
+            d="M0,1700 L0,370 L1440,0 L1440,1700 L0,1700 Z"
           ></path>
         </svg>
       </div>
