+++ styles/globals.css
@@ -11,6 +11,7 @@ body {
   padding: 0;
   margin: 0;
   font-family: 'Lexend Deca', sans-serif;
+  height: 100%;
 }
 
 a {
@@ -25,4 +26,8 @@ a {
 Image {
   min-width: none;
   min-height: none;
+}
+
+#__next {
+  height: 100%;
 }
